html/requirements.txt
django>=1.7
djangorestframework
numpy
pillow
pyyaml
scikit-learn>=0.15
sorl-thumbnail
sqlalchemy>=0.9.1
